+++ R/paged.R
@@ -5,6 +5,15 @@
 #' output document is split into multiple pages via a JavaScript library
 #' \pkg{paged.js}. These pages contain elements commonly seen in PDF documents,
 #' such as page numbers and running headers.
+#'
+#' When a path or an url is passed to the \code{front_cover} or \code{back_cover}
+#' argument, several CSS variables are created. They are named \code{--front-cover}
+#' and \code{--back-cover} and can be used as value for the CSS property \code{background-image}.
+#' For example, \code{background-image: var(--front-cover);}. When a vector of
+#' paths or urls is used as a value for \code{front_cover} or \code{back_cover},
+#' the CSS variables are suffixed with an index: \code{--front-cover-1},
+#' \code{--front-cover-2}, etc.
+#'
 #' @param ... Arguments passed to
 #'   \code{bookdown::\link[bookdown]{html_document2}}.
 #' @param css A character vector of CSS file paths. If a path does not contain
@@ -16,21 +25,27 @@
 #' @param template The path to the Pandoc template to convert Markdown to HTML.
 #' @param csl The path of the Citation Style Language (CSL) file used to format
 #'   citations and references (see the \href{https://pandoc.org/MANUAL.html#citations}{Pandoc documentation}).
+#' @param front_cover,back_cover Paths or urls to image files to be used
+#'   as front or back covers. Theses images are available through CSS variables
+#'   (see Details).
 #' @references \url{https://pagedown.rbind.io}
 #' @return An R Markdown output format.
 #' @import stats utils
 #' @export
 html_paged = function(
   ..., css = c('default-fonts', 'default-page', 'default'), theme = NULL,
-  template = pkg_resource('html', 'paged.html'), csl = NULL
+  template = pkg_resource('html', 'paged.html'), csl = NULL,
+  front_cover = NULL, back_cover = NULL
 ) {
   html_format(
     ..., css = css, theme = theme, template = template, .pagedjs = TRUE,
     .pandoc_args = c(
       lua_filters('uri-to-fn.lua', 'loft.lua', 'footnotes.lua'), # uri-to-fn.lua must come before footnotes.lua
       if (!is.null(csl)) c('--csl', csl),
-      pandoc_chapter_name_args()
-    )
+      pandoc_chapter_name_args(),
+      pandoc_covers_args(front_cover, back_cover)
+    ),
+    .dependencies = covers_dependencies(front_cover, back_cover)
   )
 }
 
@@ -166,4 +181,37 @@ pandoc_metadata_arg = function(name, value) {
 
 pandoc_chapter_name_args = function() {
   unlist(lapply(chapter_name(), pandoc_metadata_arg, name = 'chapter_name'))
+}
+
+pandoc_covers_args = function(front_cover, back_cover) {
+  build_links = function(name, img) {
+    if (length(img) == 0) return()
+    name = paste(name, seq_along(img), sep = '-')
+    links = paste0('<link id="', name,'-pagedown-attachment" rel="attachment" href="', img, '" />')
+    links[!file.exists(img)]
+  }
+  links = c(build_links('front-cover', front_cover), build_links('back-cover', back_cover))
+  if (length(links)) {
+    writeLines(links, f <- tempfile(fileext = ".html"))
+    rmarkdown::includes_to_pandoc_args(rmarkdown::includes(f))
+  }
+}
+
+covers_dependencies = function(front_cover, back_cover) {
+  html_dep = function(name, img) {
+    htmltools::htmlDependency(
+      name, packageVersion('pagedown'), dirname(path.expand(img)),
+      attachment = c(pagedown = basename(img)), all_files = FALSE
+    )
+  }
+
+  build_deps = function(name, img) {
+    if (length(img) == 0) return(list())
+    name = paste(name, seq_along(img), sep = '-')
+    name = name[file.exists(img)]
+    img = img[file.exists(img)]
+    mapply(name, img, FUN = html_dep, USE.NAMES = FALSE, SIMPLIFY = FALSE)
+  }
+
+  c(build_deps('front-cover', front_cover), build_deps('back-cover', back_cover))
 }